Add jtag programing

+++ run.sh
@@ -5,6 +5,7 @@ DEFAULT_REPO_PATH="../zmk/app"
 
 LEFT=false
 RIGHT=false
+BOOTLOADER=true
 
 POSITIONAL=()
 while [[ $# -gt 0 ]]; do
@@ -19,6 +20,10 @@ while [[ $# -gt 0 ]]; do
         RIGHT=true
         shift # past arg
         ;;
+    -n | --noboot)
+        BOOTLOADER=false
+        shift # past arg
+        ;;
     *)                     # unknown option
         POSITIONAL+=("$1") # save it in an array for later
         shift              # past argument
@@ -29,6 +34,8 @@ set -- "${POSITIONAL[@]}" # restore positional parameters
 
 (
     cd $DEFAULT_REPO_PATH
+    
+    # clean up
     rm -rf dfu-package.zip build
 
     if $LEFT; then
@@ -42,8 +49,24 @@ set -- "${POSITIONAL[@]}" # restore positional parameters
         exit 1
     fi
 
-    latestDevice=$(ls /dev/ttyACM* -1tr | tail -n1)
-    rm -rf dfu-package.zip
-    adafruit-nrfutil dfu genpkg --dev-type 0x0052 --application build/zephyr/*.hex dfu-package.zip
-    adafruit-nrfutil dfu serial --package dfu-package.zip -p "$latestDevice" -b 115200
+
+    if $BOOTLOADER; then
+        FLASH_IMG=$(find build/zephyr -name "dactyl_manuform*.hex")
+        echo "Flashing $FLASH_IMG"
+
+        latestDevice=$(ls /dev/ttyACM* -1tr | tail -n1)
+        rm -rf dfu-package.zip
+        adafruit-nrfutil dfu genpkg --dev-type 0x0052 --application "$FLASH_IMG" dfu-package.zip
+        adafruit-nrfutil dfu serial --package dfu-package.zip -p "$latestDevice" -b 115200
+    else
+        FLASH_IMG=$(find build/zephyr -name "dactyl_manuform*.elf")
+        echo "Flashing $FLASH_IMG"
+
+        openocd -f interface/ftdi/jtag-lock-pick_tiny_2.cfg \
+                -c "transport select swd" \
+                -f target/nrf52.cfg \
+                -c "init" \
+                -c "program $FLASH_IMG verify" \
+                -c "exit"
+    fi
 )
